创建订单，单元测试

+++ spring-boot-weixin-sell/src/main/java/com/amarsoft/springboot/VO/ProductInfoVO.java
@@ -31,7 +31,7 @@ public class ProductInfoVO {
 	private String productName;
 	
 	@JsonProperty("price")
-	private double productPrice;
+	private BigDecimal productPrice;
 	
 	@JsonProperty("description")
 	private String ProductDescription;
